Rebar3 and Thumbs, but major other changes as well.
Previously, this package built a port executable.
That version had no unit tests, and upon adding some simple ones it became apparent that it didn't work as expected.
For efficiency reasons, it seemed desirable to implement it as a NIF instead of a port, so this commit does that, as well as adding tests and docs.
This is a breaking change for riak_auth_mods, which must also use this branch.